requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
coremltools==6.1
filelock==3.9.0
huggingface-hub==0.11.1
idna==3.4
mpmath==1.2.1
numpy==1.23.0
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
PyYAML==6.0
requests==2.28.1
sympy==1.11.1
timm==0.6.12
torchprofile==0.0.4
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13